requirements.txt
web3==6.11.1
python-dotenv==1.0.0
py-solc-x==2.0.1
cryptography==41.0.5
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.0
hexbytes==0.3.1
